Add config options
* defaultTheme
* defaultSize (TODO)
* includedThemes
* excludedThemes

exclude will take out themes from include, and if include is blank, all themes are included, with
exclude removing from there as well.

If you exclude a theme, and then use it, the component will still render,
but will be transparent. The todo here is to use 'default', which probably
should always be included.